cleanup: file structure
